# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/curd/functionaltemplate/add": {
            "post": {
                "description": "新增",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "功能模版管理"
                ],
                "summary": "新增",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add FunctionalTemplate",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.FunctionalTemplateInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/functionaltemplate/all": {
            "get": {
                "description": "查询所有",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "功能模版管理"
                ],
                "summary": "查询所有",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.GetAllFunctionalTemplateResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/functionaltemplate/copy": {
            "post": {
                "description": "复制",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "功能模版管理"
                ],
                "summary": "复制",
                "parameters": [
                    {
                        "description": "Copy FunctionalTemplate",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.DelRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/functionaltemplate/delete": {
            "delete": {
                "description": "删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "功能模版管理"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Delete FunctionalTemplate",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.DelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/functionaltemplate/detail": {
            "get": {
                "description": "查询明细",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "功能模版管理"
                ],
                "summary": "查询明细",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.GetFunctionalTemplateDetailResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/functionaltemplate/enable": {
            "post": {
                "description": "禁用/启用",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "功能模版管理"
                ],
                "summary": "禁用/启用",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Enable/Disable FunctionalTemplate",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.EnableRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/functionaltemplate/export": {
            "get": {
                "description": "导出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "功能模版管理"
                ],
                "summary": "导出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "从1开始",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "默认每页10条",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "orderField",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否倒序排序",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "语音",
                        "name": "language",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分组",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "IDs",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/functionaltemplate/import": {
            "post": {
                "description": "导入",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "功能模版管理"
                ],
                "summary": "导入",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer+空格+Token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "要上传的文件",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/functionaltemplate/query": {
            "get": {
                "description": "分页查询",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "功能模版管理"
                ],
                "summary": "分页查询",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "从1开始",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "默认每页10条",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "orderField",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否倒序排序",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "语音",
                        "name": "language",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "分组",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.QueryFunctionalTemplateResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/functionaltemplate/update": {
            "put": {
                "description": "更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "功能模版管理"
                ],
                "summary": "更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update FunctionalTemplate",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.FunctionalTemplateInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
...
        "/api/curd/template/copy": {
            "post": {
                "description": "复制",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "代码模板管理"
                ],
                "summary": "复制",
                "parameters": [
                    {
                        "description": "Copy Template",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.DelRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "curd.CommonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "curd.DelRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "tenantID": {
                    "type": "string"
                }
            }
        },
        "curd.EnableRequest": {
            "type": "object",
            "properties": {
                "enable": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "tenantID": {
                    "type": "string"
                }
            }
        },
...
        "curd.FunctionalTemplateInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "fileTemplateIDs": {
                    "description": "文件模版ID",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "group": {
                    "description": "分组",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "description": "语音",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "params": {
                    "description": "自定义参数",
                    "type": "string"
                }
            }
        },
...
        "curd.GetAllFunctionalTemplateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/curd.FunctionalTemplateInfo"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "curd.GetFunctionalTemplateDetailResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/curd.FunctionalTemplateInfo"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "curd.QueryFunctionalTemplateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/curd.FunctionalTemplateInfo"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pages": {
                    "type": "integer"
                },
                "records": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}